Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/svg/animations/end-attribute-change-end-time-expected.txt


FAIL Mutation of the 'end' attribute changes current interval end assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
